Finished a Basic Audio Settings Page
In the future we plan on using the volume mixer to make it easier to put sounds in different mixers. So that players can turn off the music while still having the sound effects of the game on max.

However. Due to myself not being familiar with the Volume Mixer. For now, this will do and can setup the basis of a full colume setup